# docs/openapi.yaml
openapi: 3.0.3
info:
  title: IT-Ticket API
  version: 1.1.0
  description: >
    REST-API för IT-Ticket — ett ärendehanteringssystem för intern IT-support på Prefabmästarna.


    **Leverans:** rullande från `main` (en instans per deployment).
    Versionen ovan spårar API-kontraktet, inte appens git-tagg.


    ## API coverage

    This spec documents the full API surface: 160 operations
    across every route file — health/CSRF, auth, tickets, categories,
    contacts, companies, kb, public, users, API keys,
    settings, comments, attachments, shares, links, checklists, checklist
    templates, template fields, tags,
    reports, backup, push notifications, webhooks, and inbound email.


    **Prose reference:** `docs/API.md` — generated from the actual Express
    routers in `server/src/routes/` and kept up to date on route changes. Use
    it as a narrative cross-check alongside this machine-readable spec.


    ## Autentisering

    Två sätt att autentisera, båda via `Authorization`-headern:

    - **Session (JWT):** `Authorization: Bearer <jwt>`. Access-token gäller 15 min;
      en HttpOnly refresh-cookie (path `/api/auth`) förnyar den via
      `POST /api/auth/refresh`. **Muterande anrop (POST/PUT/PATCH/DELETE) kräver
      även CSRF-token** i `x-csrf-token` (hämtas från `GET /api/csrf-token`).
    - **API-nyckel:** `Authorization: Bearer itk_live_...`. Nycklar har scope
      `read`, `read,write` eller — för administrativa endpoints — även `admin`.
...


...

...

...

...

...

...

...

...
servers:
...
security:
  - bearerAuth: []
  - apiKeyAuth: []
paths:
...
  /tickets:
    get:
      tags: [Tickets]
      summary: Lista ärenden
      description: >
        Utan `page`/`limit` returneras en ren array (max 1000, senaste först,
        utan `description`/`notes`/`solution`). Med `page` eller `limit`
        returneras `{ data, pagination }`. `countOnly=true` returnerar bara
        `{ count }` för de aktuella filtren.
      parameters:
        - { name: page, in: query, schema: { type: integer, minimum: 1 } }
        - { name: limit, in: query, schema: { type: integer, enum: [10, 20, 25, 30, 50, 100, 1000] }, description: "Andra värden faller tillbaka på 10." }
        - { name: status, in: query, schema: { type: string }, description: "En eller flera statusar, kommaseparerade (open, in-progress, waiting, resolved, closed)." }
        - { name: priority, in: query, schema: { type: string } }
        - { name: category, in: query, schema: { type: string }, description: Kategori-id. }
        - { name: company_id, in: query, schema: { type: string } }
        - { name: assigned_to, in: query, schema: { type: string } }
        - { name: requester_id, in: query, schema: { type: string } }
        - { name: search, in: query, schema: { type: string }, description: Fulltextsökning (FTS5) }
        - { name: tags, in: query, schema: { type: string }, description: Kommaseparerade tagg-id. }
        - { name: tagMode, in: query, schema: { type: string, enum: [any, all] } }
        - { name: dateFrom, in: query, schema: { type: string, format: date } }
        - { name: dateTo, in: query, schema: { type: string, format: date } }
        - { name: dateField, in: query, schema: { type: string } }
        - { name: checklist, in: query, schema: { type: string } }
        - { name: year, in: query, schema: { type: integer } }
        - { name: month, in: query, schema: { type: integer, minimum: 0, maximum: 11 } }
        - { name: sortBy, in: query, schema: { type: string, enum: [createdAt, status, priority, category], default: createdAt } }
        - { name: sortDir, in: query, schema: { type: string, enum: [asc, desc], default: desc } }
        - { name: countOnly, in: query, schema: { type: boolean } }
      responses:
        "200":
          description: Array (utan paginering), sidindelat svar eller `{ count }`
          content:
            application/json:
              schema:
                oneOf:
                  - type: array
                    items: { $ref: "#/components/schemas/TicketListItem" }
                  - type: object
                    properties:
                      data:
                        type: array
                        items: { $ref: "#/components/schemas/TicketListItem" }
                      pagination: { $ref: "#/components/schemas/Pagination" }
                  - type: object
                    properties:
                      count: { type: integer }
        "401": { $ref: "#/components/responses/Unauthorized" }
    post:
      tags: [Tickets]
      summary: Skapa ärende
      security:
        - bearerAuth: []
          csrfToken: []
        - apiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TicketInput" }
      responses:
        "201":
          description: Skapat ärende
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Ticket" }
        "400": { $ref: "#/components/responses/Error" }
...
  /categories:
    get:
      tags: [Categories]
      summary: Lista kategorier
      responses:
        "200":
          description: Kategorier
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Category" }
    post:
      tags: [Categories]
      summary: Skapa kategori (admin)
      security:
        - bearerAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [label]
              properties:
                label: { type: string }
      responses:
        "201":
          description: Skapad kategori
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /public/tickets:
    post:
      tags: [Public]
      summary: Skapa ärende från portalen
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, title]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                title: { type: string }
                description: { type: string }
                category: { type: string }
                priority: { type: string, enum: [low, medium, high, critical] }
                template_id: { type: string }
                website:
                  type: string
                  description: >
                    Honeypot - ska vara tomt (dolt fält i formuläret). Ett ifyllt värde ger
                    ett falskt `200` med slumpat `ticketId` och ingenting sparas.
                formStartedAt:
                  description: Epoch-ms eller ISO-tid då formuläret öppnades. Inskick snabbare än en minsta fyllnadstid avvisas med `400`.
                  oneOf:
                    - { type: integer }
                    - { type: string, format: date-time }
                customFields:
                  type: array
                  items:
                    type: object
                    properties:
                      fieldName: { type: string }
                      fieldLabel: { type: string }
                      fieldValue: { type: string }
      responses:
        "200":
          description: Honeypot utlöst (inget ärende skapades)
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  ticketId: { type: string }
        "201":
          description: Skapat ärende
          content:
            application/json:
              schema:
                type: object
                properties:
                  message: { type: string }
                  ticketId: { type: string }
        "400": { $ref: "#/components/responses/Error" }
...
  /templates:
    get:
      operationId: templatesList
      tags: [Templates]
      summary: List ticket templates
      description: Any authenticated user. Returns every ticket template with its dynamic fields (empty array for standard templates).
      responses:
        "200":
          description: Ticket templates
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Template" }
    post:
      operationId: templatesCreate
      tags: [Templates]
      summary: Create a ticket template (admin)
      description: >
        Admin only. `description_template` is required unless `template_type`
        is `dynamic`. When `template_type` is `dynamic`, `fields` (if given)
        is inserted inline as the template's fields in the same request —
        there is no need to call the template-fields endpoints separately at
        creation time.
      security:
        - bearerAuth: []
          csrfToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/TemplateInput" }
      responses:
        "201":
          description: Created template with fields
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Template" }
        "400": { $ref: "#/components/responses/Error" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
  /public/branding/logo:
    get:
      operationId: publicBrandingLogo
      tags: [Public]
      summary: Instance logo bytes
      description: >
        PNG, JPEG or WebP (SVG is not accepted). Served `inline` with
        `X-Content-Type-Options: nosniff` and `Cross-Origin-Resource-Policy:
        cross-origin` so it can render on the login page and public form.
      security: []
      responses:
        "200":
          description: Image
          content:
            image/png: { schema: { type: string, format: binary } }
            image/jpeg: { schema: { type: string, format: binary } }
            image/webp: { schema: { type: string, format: binary } }
        "404":
          description: No logo configured
          $ref: "#/components/responses/Error"

  # ───────────────────────── Settings: branding logo ─────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Session-JWT: `Authorization: Bearer <jwt>`."
    apiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "API-nyckel: `Authorization: Bearer itk_live_...` (scope read/write/admin; admin-skyddade endpoints kräver `admin` även när ägaren är administratör). Undantagen från CSRF."
    csrfToken:
      type: apiKey
      in: header
      name: x-csrf-token
      description: Krävs på muterande sessions-anrop. Hämtas från `GET /api/csrf-token`.
  responses:
    Error:
      description: Fel
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
    Unauthorized:
      description: Ej autentiserad
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
    Forbidden:
      description: Saknar behörighet (t.ex. admin krävs eller fel scope)
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      properties:
        error: { type: string }
      required: [error]
...
    Ticket:
      type: object
      properties:
        id: { type: string }
        title: { type: string }
        description: { type: string }
        status: { type: string, enum: [open, in-progress, waiting, resolved, closed] }
        priority: { type: string, enum: [low, medium, high, critical] }
        category_id: { type: string, nullable: true }
        requester_id: { type: string, nullable: true }
        requester_name: { type: string, nullable: true }
        category_label: { type: string, nullable: true }
        company_id: { type: string, nullable: true }
        assigned_to: { type: string, nullable: true }
        assigned_to_name: { type: string, nullable: true }
        notes: { type: string, nullable: true }
        solution: { type: string, nullable: true }
        created_by: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        resolved_at: { type: string, format: date-time, nullable: true }
        closed_at: { type: string, format: date-time, nullable: true }
    TicketInput:
      type: object
      required: [title]
      description: Antingen `description` eller `customFields` krävs.
      properties:
        title: { type: string }
        description: { type: string }
        status: { type: string, enum: [open, in-progress, waiting, resolved, closed] }
        priority: { type: string, enum: [low, medium, high, critical] }
        category_id: { type: string }
        requester_id: { type: string }
        company_id: { type: string }
        assigned_to: { type: string }
        notes: { type: string }
        solution: { type: string }
        template_id: { type: string }
        customFields:
          type: array
          items:
            type: object
            properties:
              fieldName: { type: string }
              fieldLabel: { type: string }
              fieldValue: { type: string }
    Category:
      type: object
      properties:
        id: { type: string }
        name: { type: string }
        label: { type: string }
        position: { type: integer }
        created_at: { type: string, format: date-time }
...
    Template:
      type: object
      description: >
        A reusable ticket blueprint used to prefill `POST /tickets`. `standard`
        templates prefill title/description/etc.; `dynamic` templates instead
        carry a set of custom form `fields` used to compose the ticket body.
      properties:
        id: { type: string }
        name: { type: string }
        description: { type: string, nullable: true }
        template_type: { type: string, enum: [standard, dynamic] }
        title_template: { type: string }
        description_template: { type: string, description: Empty string for dynamic templates rather than null (column is NOT NULL). }
        priority: { type: string, enum: [low, medium, high, critical] }
        category_id: { type: string, nullable: true }
        notes_template: { type: string, nullable: true }
        solution_template: { type: string, nullable: true }
        position: { type: integer }
        created_by: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        fields:
          type: array
          items: { $ref: "#/components/schemas/TemplateField" }
...
    Pagination:
      type: object
      required: [page, limit, total]
      properties:
        page: { type: integer }
        limit: { type: integer }
        total: { type: integer }
        totalPages: { type: integer, description: Only on ticket lists. }
        hasNext: { type: boolean, description: Only on ticket lists. }
        hasPrev: { type: boolean, description: Only on ticket lists. }
    TicketListItem:
      type: object
      description: Ticket as returned by list endpoints - without the heavy `description`, `notes` and `solution` fields.
      properties:
        id: { type: string }
        title: { type: string }
        status: { type: string, enum: [open, in-progress, waiting, resolved, closed] }
        priority: { type: string, enum: [low, medium, high, critical] }
        category_id: { type: string, nullable: true }
        category_label: { type: string, nullable: true }
        requester_id: { type: string, nullable: true }
        requester_name: { type: string, nullable: true }
        company_id: { type: string, nullable: true }
        assigned_to: { type: string, nullable: true }
        assigned_to_name: { type: string, nullable: true }
        template_id: { type: string, nullable: true }
        created_at: { type: string, format: date-time }
        updated_at: { type: string, format: date-time }
        resolved_at: { type: string, format: date-time, nullable: true }
        closed_at: { type: string, format: date-time, nullable: true }
...